requirements.txt
google-genai
opencv-python
pillow
mss
librosa 
soundfile
python-dotenv
openai
pathlib
python-socketio[asyncio-client] 
pygame 
aiohttp
fastapi
numpy
openai[voice_helpers]
sounddevice
uvicorn 
python-multipart
httpx
pandas
#
office365-rest-python-client 
openpyxl 
joblib
